nebula-backend/requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
thirdweb-ai>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
